refactor(express): remove renderLocations hook
...change `configureServer` hook strategy to `sequence` and move
irrelevant settings to `@untool/webpack` preset
