+++ packages/express/preset.js
@@ -6,9 +6,6 @@ module.exports = {
   https: false,
   host: '[HOST]',
   port: '[PORT]',
-  locations: [],
-  basePath: '',
-  assetPath: '<basePath>',
-  buildDir: join('<rootDir>', 'dist'),
+  distDir: join('<rootDir>', 'dist'),
   mixins: [__dirname],
 };
